separated the logic from main.js by components
